requirements.txt
asgiref==3.8.1
awscli==1.33.18
blinker==1.8.2
botocore==1.34.136
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
Django==4.2.3
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.16
flask==3.0.3
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
mako==1.3.5
markdown==3.2.2
markupsafe==2.1.5
numpy==1.26.4
pathspec==0.10.1
pdoc3==0.9.1
pillow==10.3.0
postgis==1.0.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyjwt==2.8.0
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
redis>=4.0.0
# Include pywin32 conditionally
pywin32==306; sys_platform == 'win32'
requests==2.25.1
celery==5.2.3
django-cors-headers